requirements.txt
openai
python-dotenv
langchain-openai